requirements.txt
# PyTorch 및 관련 라이브러리
torch>=2.5.0,<2.6.0
torchvision>=0.20.0,<0.21.0
pytorch-lightning>=2.4.0,<2.5.0
torchmetrics>=1.4.0,<1.5.0

# 모델 및 아키텍처
timm>=1.0.0,<1.1.0

# Hydra (설정 관리)
hydra-core>=1.3.0,<1.4.0
omegaconf>=2.3.0,<2.4.0

# 데이터 처리
pandas>=2.2.0,<2.3.0
numpy>=1.26.0,<2.0.0
opencv-python-headless>=4.10.0,<4.11.0
albumentations >=2.0.0,<3.0.0
pydantic>=2.10.0,<3.0.0
Pillow>=10.4.0,<11.0.0

# 시각화 처리
matplotlib>=3.8.0,<3.11.0
seaborn>=0.13.0,<0.14.0

# 메트릭 및 평가
scikit-learn>=1.5.0,<1.6.0

# 실험 추적
wandb>=0.18.0,<0.19.0
tensorboard>=2.17.0,<2.18.0

# 유틸리티
tqdm>=4.66.0,<4.67.0
pyyaml>=6.0.0,<7.0.0
rich>=13.7.0,<14.0.0
python-dotenv>=1.0.0,<2.0.0

# 코드 품질 (선택사항)
black>=24.0.0,<25.0.0
# flake8>=7.0.0,<8.0.0
# mypy>=1.11.0,<1.12.0
